feat: split createServer and index.ts
